+++ game/player.py
@@ -1,0 +1,6 @@
+class Player:
+    def __init__(self):
+        self.name = None
+        self.health = 100
+        self.damage = 15
+        self.coins = 0
